alterando add e edit
