--- HortCult/HortCult/Views/FormPlant/Components/Buttons/ButtonPattern.swift
@@ -1,20 +1,0 @@
-//
-//  ButtonPattern.swift
-//  HortCult
-//
-//  Created by Joao Guilherme Araujo Canuto on 05/07/23.
-//
-
-import SwiftUI
-
-struct ButtonPattern: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct ButtonPattern_Previews: PreviewProvider {
-    static var previews: some View {
-        ButtonPattern()
-    }
-}
