+++ HortCult/HortCult/Views/FormPlant/Components/Pickers/CategoryPicker.swift
@@ -1,20 +1,20 @@
 //
-//  EditButton.swift
+//  CategoryPicker.swift
 //  HortCult
 //
-//  Created by Joao Guilherme Araujo Canuto on 05/07/23.
+//  Created by Joao Guilherme Araujo Canuto on 06/07/23.
 //
 
 import SwiftUI
 
-struct EditButton: View {
+struct CategoryPicker: View {
     var body: some View {
         Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
     }
 }
 
-struct EditButton_Previews: PreviewProvider {
+struct CategoryPicker_Previews: PreviewProvider {
     static var previews: some View {
-        EditButton()
+        CategoryPicker()
     }
 }
