+++ HortCult/HortCult/Views/FormPlant/Components/Pickers/FrequencyPicker.swift
@@ -1,20 +1,20 @@
 //
-//  DeleteButton.swift
+//  FrequencyPicker.swift
 //  HortCult
 //
-//  Created by Joao Guilherme Araujo Canuto on 05/07/23.
+//  Created by Joao Guilherme Araujo Canuto on 06/07/23.
 //
 
 import SwiftUI
 
-struct DeleteButton: View {
+struct FrequencyPicker: View {
     var body: some View {
         Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
     }
 }
 
-struct DeleteButton_Previews: PreviewProvider {
+struct FrequencyPicker_Previews: PreviewProvider {
     static var previews: some View {
-        DeleteButton()
+        FrequencyPicker()
     }
 }
