+++ HortCult/HortCult/Views/FormPlant/Components/TextFields/DescriptionInput.swift
@@ -1,0 +1,20 @@
+//
+//  DescriptionInput.swift
+//  HortCult
+//
+//  Created by Joao Guilherme Araujo Canuto on 06/07/23.
+//
+
+import SwiftUI
+
+struct DescriptionInput: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct DescriptionInput_Previews: PreviewProvider {
+    static var previews: some View {
+        DescriptionInput()
+    }
+}
