+++ HortCult/HortCult/Views/FormPlant/Components/TextFields/NameInput.swift
@@ -1,20 +1,20 @@
 //
-//  AddButton.swift
+//  NameInput.swift
 //  HortCult
 //
-//  Created by Joao Guilherme Araujo Canuto on 05/07/23.
+//  Created by Joao Guilherme Araujo Canuto on 06/07/23.
 //
 
 import SwiftUI
 
-struct AddButton: View {
+struct NameInput: View {
     var body: some View {
         Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
     }
 }
 
-struct AddButton_Previews: PreviewProvider {
+struct NameInput_Previews: PreviewProvider {
     static var previews: some View {
-        AddButton()
+        NameInput()
     }
 }
