+++ HortCult/HortCult/Views/InformationScreen/InformationViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  InformationViewModel.swift
+//  HortCult
+//
+//  Created by Joao Guilherme Araujo Canuto on 06/07/23.
+//
+
+import Foundation
